change protein, carbohydrates and fat to FloatField

+++ backend/products_and_meals/api/serializers.py
@@ -4,9 +4,9 @@ from django.contrib.auth.models import User
 from django.core.validators import MaxValueValidator, MinValueValidator
 
 class ProductSerializer(serializers.ModelSerializer):
-    protein = serializers.IntegerField(validators=[MaxValueValidator(100), MinValueValidator(0)])
-    carbohydrates = serializers.IntegerField(validators=[MaxValueValidator(100), MinValueValidator(0)])
-    fat = serializers.IntegerField(validators=[MaxValueValidator(100), MinValueValidator(0)])
+    protein = serializers.FloatField(validators=[MaxValueValidator(100), MinValueValidator(0)])
+    carbohydrates = serializers.FloatField(validators=[MaxValueValidator(100), MinValueValidator(0)])
+    fat = serializers.FloatField(validators=[MaxValueValidator(100), MinValueValidator(0)])
 
     class Meta:
         model = Product
@@ -18,9 +18,9 @@ class SummarySerializer(serializers.ModelSerializer):
         exclude = ['summary_id', 'user_id', 'date']
 
 class DemandSerializer(serializers.ModelSerializer):
-    protein = serializers.IntegerField(validators=[MaxValueValidator(500), MinValueValidator(0)])
-    carbohydrates = serializers.IntegerField(validators=[MaxValueValidator(2000), MinValueValidator(0)])
-    fat = serializers.IntegerField(validators=[MaxValueValidator(250), MinValueValidator(0)])
+    protein = serializers.FloatField(validators=[MaxValueValidator(500), MinValueValidator(0)])
+    carbohydrates = serializers.FloatField(validators=[MaxValueValidator(2000), MinValueValidator(0)])
+    fat = serializers.FloatField(validators=[MaxValueValidator(250), MinValueValidator(0)])
 
     class Meta:
         model = Demand
